requirements.txt
Django>=3.2,<3.3
Pillow>=8.1.2,<8.2.0
django-bootstrap-v5>=1.0.4,<1.1.0
django-widget-tweaks>=1.4.8,<1.5.0
django-braces>=1.14.0,<1.15.0
django-guardian>=2.4.0,<2.5.0
django-crispy-forms>=1.12.0,<1.13.0
crispy-bootstrap5>=0.4,<0.5
opencv-python
django-imagekit>=4.0.2,<4.1.0
channels==3.0.4
channels-redis==3.3.0
flake8>=3.9.1,<4.0.0